Resolucion de examen inlcuye Refactor, Diagrama de Clases, Secuencia y Componentes
