+++ src/main/java/library/model/Book.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.library.model;
 
 public class Book {
     private String title;
