+++ src/main/java/library/service/LibraryService.java
@@ -1,38 +1,37 @@
-package main.java;
+package main.java.library.service;
+
+import main.java.library.model.Book;
+import main.java.library.model.User;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class Library {
+public class LibraryService {
     private List<Book> books;
     private List<User> users;
 
-    public Library() {
+    public LibraryService() {
         this.books = new ArrayList<>();
         this.users = new ArrayList<>();
     }
 
-    // Agregar un libro a la biblioteca
     public void addBook(Book book) {
         if (!books.contains(book)) {
             books.add(book);
         }
     }
 
-    // Remover un libro de la biblioteca
     public void removeBook(Book book) {
         books.remove(book);
     }
 
-    // Registrar un usuario en la biblioteca
     public void registerUser(User user) {
         if (!users.contains(user)) {
             users.add(user);
         }
     }
 
-    // Permitir que un usuario tome un libro prestado
     public boolean borrowBook(User user, Book book) {
         if (books.contains(book)) {
             if (user.getType().equals("Student") && user.getBorrowedBooks().size() < 2) {
@@ -48,7 +47,6 @@ public class Library {
         return false;
     }
 
-    // Permitir que un usuario devuelva un libro
     public void returnBook(User user, Book book) {
         if (!books.contains(book)) {
             user.returnBook(book);
@@ -56,7 +54,6 @@ public class Library {
         }
     }
 
-    // Generar un reporte de la biblioteca
     public String generateReport() {
         StringBuilder report = new StringBuilder("Library Report\n");
         report.append("Available Books:\n");
@@ -72,17 +69,14 @@ public class Library {
         return report.toString();
     }
 
-    // Método para verificar si un libro está disponible en la biblioteca
     public boolean isBookAvailable(Book book) {
         return books.contains(book);
     }
 
-    // Obtener una lista inmutable de los libros
     public List<Book> getBooks() {
         return Collections.unmodifiableList(books);
     }
 
-    // Obtener una lista inmutable de los usuarios
     public List<User> getUsers() {
         return Collections.unmodifiableList(users);
     }
