+++ src/test/java/LibraryTest.java
@@ -1,15 +1,15 @@
 package test.java;
 
-import main.java.Book;
-import main.java.Library;
-import main.java.User;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
+
+import main.java.library.service.LibraryService;
+import main.java.library.model.Book;
+import main.java.library.model.User;
 
 public class LibraryTest {
 
-    private Library library;
+    private LibraryService library;
     private Book book1;
     private Book book2;
     private User student;
@@ -17,8 +17,7 @@ public class LibraryTest {
 
     @BeforeEach
     public void setUp() {
-        // Inicializar la biblioteca y objetos de prueba
-        library = new Library();
+        library = new LibraryService();
         book1 = new Book("The Catcher in the Rye");
         book2 = new Book("1984");
         student = new User("Alice", "Student");
@@ -34,28 +33,28 @@ public class LibraryTest {
     public void testAddBook() {
         Book book3 = new Book("To Kill a Mockingbird");
         library.addBook(book3);
-        assertTrue(library.getBooks().contains(book3), "La biblioteca debería contener el libro agregado");
+        assertTrue(library.getBooks().contains(book3));
     }
 
     @Test
     public void testRemoveBook() {
         library.removeBook(book1);
-        assertFalse(library.getBooks().contains(book1), "La biblioteca no debería contener el libro eliminado");
+        assertFalse(library.getBooks().contains(book1));
     }
 
     @Test
     public void testRegisterUser() {
         User newUser = new User("Charlie", "Student");
         library.registerUser(newUser);
-        assertTrue(library.getUsers().contains(newUser), "La biblioteca debería contener el usuario registrado");
+        assertTrue(library.getUsers().contains(newUser));
     }
 
     @Test
     public void testBorrowBookForStudent() {
         boolean borrowed = library.borrowBook(student, book1);
-        assertTrue(borrowed, "El estudiante debería poder tomar prestado el libro");
-        assertFalse(library.getBooks().contains(book1), "El libro prestado no debería estar disponible en la biblioteca");
-        assertTrue(student.getBorrowedBooks().contains(book1), "El libro debería aparecer en la lista de libros prestados del estudiante");
+        assertTrue(borrowed);
+        assertFalse(library.getBooks().contains(book1));
+        assertTrue(student.getBorrowedBooks().contains(book1));
     }
 
     @Test
@@ -66,32 +65,32 @@ public class LibraryTest {
         library.addBook(book3);
 
         boolean borrowed = library.borrowBook(student, book3);
-        assertFalse(borrowed, "El estudiante no debería poder tomar prestado más libros que el límite permitido");
+        assertFalse(borrowed);
     }
 
     @Test
     public void testBorrowBookForTeacher() {
         boolean borrowed = library.borrowBook(teacher, book1);
-        assertTrue(borrowed, "El profesor debería poder tomar prestado el libro");
-        assertFalse(library.getBooks().contains(book1), "El libro prestado no debería estar disponible en la biblioteca");
-        assertTrue(teacher.getBorrowedBooks().contains(book1), "El libro debería aparecer en la lista de libros prestados del profesor");
+        assertTrue(borrowed);
+        assertFalse(library.getBooks().contains(book1));
+        assertTrue(teacher.getBorrowedBooks().contains(book1));
     }
 
     @Test
     public void testReturnBook() {
         library.borrowBook(student, book1);
         library.returnBook(student, book1);
-        assertTrue(library.getBooks().contains(book1), "El libro devuelto debería estar disponible en la biblioteca");
-        assertFalse(student.getBorrowedBooks().contains(book1), "El libro devuelto no debería estar en la lista de libros prestados del estudiante");
+        assertTrue(library.getBooks().contains(book1));
+        assertFalse(student.getBorrowedBooks().contains(book1));
     }
 
     @Test
     public void testGenerateReport() {
         library.borrowBook(student, book1);
         String report = library.generateReport();
-        assertTrue(report.contains("Available Books:"), "El reporte debería contener la sección de libros disponibles");
-        assertTrue(report.contains("Borrowed Books:"), "El reporte debería contener la sección de libros prestados");
-        assertTrue(report.contains("1984"), "El reporte debería listar el libro '1984' como disponible");
-        assertTrue(report.contains("The Catcher in the Rye borrowed by Alice"), "El reporte debería mostrar que 'The Catcher in the Rye' fue tomado prestado por Alice");
+        assertTrue(report.contains("Available Books:"));
+        assertTrue(report.contains("Borrowed Books:"));
+        assertTrue(report.contains("1984"));
+        assertTrue(report.contains("The Catcher in the Rye borrowed by Alice"));
     }
 }
